+++ tingting/src/main/java/com/github/longkai/tx/repo/ResourceRepo.java
@@ -32,7 +32,7 @@ public interface ResourceRepo extends BaseRepo<Resource> {
 	 * @param count 多少条呢
 	 * @return my uploaded resources
 	 */
-	Resource[] myResources(@Param("tid") long teacherId, @Param("offset") int offset, @Param("count") int count);
+	Resource[] myResources(@Param("tid") long teacherId, @Param("offset") long offset, @Param("count") int count);
 
 	/**
 	 * 某个课程下的资源
@@ -41,7 +41,7 @@ public interface ResourceRepo extends BaseRepo<Resource> {
 	 * @param count 数量
 	 * @return the resources from the course
 	 */
-	Resource[] resourcesFromCourse(@Param("cid") long cid, @Param("offset") int offset, @Param("count") int count);
+	Resource[] resourcesFromCourse(@Param("cid") long cid, @Param("offset") long offset, @Param("count") int count);
 
 	/**
 	 * 搜索资源（注意，懒加载，只返回资源相关的信息）
@@ -50,5 +50,5 @@ public interface ResourceRepo extends BaseRepo<Resource> {
 	 * @param count 多少条
 	 * @return search resources
 	 */
-	Resource[] searchResources(@Param("keywords") String keywords, @Param("offset") int offset, @Param("count") int count);
+	Resource[] searchResources(@Param("keywords") String keywords, @Param("offset") long offset, @Param("count") int count);
 }
